Attempt to make G-Java work better.
git-svn-id: https://svn.code.sf.net/p/gjava/code@879 0bf6c8e7-452b-0410-8c87-9aaedaf074f6

+++ G-Java/src/org/gcreator/compilers/gjava/api/Variables.java
@@ -557,7 +557,7 @@ public static Object   argument
     }
 
     public Object getFps() {
-        return new Integer(Game.Current.getFPS());
+        return new Integer((int) Game.Current.getFPS());
     }
 
     public void setFps(Object fps) {
@@ -731,7 +731,7 @@ public static Object   argument
     }
 
     public void setRoom_speed(Object room_speed) {
-        Game.Current.speed = room_speed.getDouble();
+        Game.Current.speed = (long) room_speed.getDouble();
     }
 
     public Object getRoom_width() {
